Fix #41 : modify the function place according to semantics file name
